+++ app/Commands/ReleaseCommand.php
@@ -63,8 +63,7 @@ class ReleaseCommand extends Command
     public function handle()
     {
         if ( ! $this->dir->hasChanges()) {
-            $this->build('No changes.');
-            $this->info("Changelog for {$this->argument('tag')} created");
+           $this->info("No Changes -> No Changelog for {$this->argument('tag')} created");
            return;
         }
 
@@ -129,15 +128,24 @@ CONTENT;
             $count   = $logType->count();
             $changes = sprintf('%d %s', $count, $count === 1 ? 'change' : 'changes');
             $content = "### {$header} ({$changes})\n\n";
+            $markdownOptions = $this->config->getMarkdownOptions();
 
             if ($this->config->hasGroups()) {
                 $content .= $logType->sort(function (Collection $logA,  Collection $logB) {
                     return $this->config->compare($logA->first()->group(), $logB->first()->group());
-                })->map(static function (Collection $group, $name) {
-                    $content = "#### {$name}\n\n";
+                })->map(static function (Collection $group, $name) use ($markdownOptions){
+                    if ($markdownOptions['groupsAsList']) {
+                        $content = "{$markdownOptions['listStyle']} **{$name}**\n";
+                    } else {
+                        $content = "#### {$name}\n\n";
+                    }
 
-                    $content .= $group->map(static function (LogEntry $log) {
-                        $changeEntry = "- {$log->title()}";
+                    $content .= $group->map(static function (LogEntry $log) use ($markdownOptions) {
+                        $changeEntry = "";
+                        if ($markdownOptions['groupsAsList']) {
+                            $changeEntry = "  ";
+                        }
+                        $changeEntry .= "{$markdownOptions['listStyle']} {$log->title()}";
 
                         if ($log->hasAuthor()) {
                             $changeEntry .= " (props {$log->author()})";
@@ -149,8 +157,8 @@ CONTENT;
                     return $content;
                 })->implode("\n\n");
             } else {
-                $content .= $logType->map(static function (LogEntry $log) {
-                    $changeEntry = "- {$log->title()}";
+                $content .= $logType->map(static function (LogEntry $log) use ($markdownOptions) {
+                    $changeEntry = "{$markdownOptions['listStyle']} {$log->title()}";
 
                     if ($log->hasAuthor()) {
                         $changeEntry .= " (props {$log->author()})";