+++ config/changelogger.php
@@ -25,4 +25,8 @@ return [
             'Other'                   => 'other',
             'No Changelog'            => 'ignore',
     ],
+    'markdown'      => [
+        'listStyle' => '-',
+        'groupsAsList' => false,
+    ],
 ];